fix(PR-393): extend transient effect checks to CantWinTheGame, CantGainLife, CantLoseLife
The Everybody Lives! fix in sba.rs correctly added transient_continuous_effects
scanning to player_has_cant_lose, but the sibling queries player_has_cant_win,
player_has_cant_gain_life, and player_has_cant_lose_life in static_abilities.rs
only scanned battlefield permanents via check_static_ability.

Add pub(crate) transient_grants_static_mode_to_player to static_abilities.rs
(the canonical home for player-scoped protection queries) and update all three
sibling functions to OR in the transient scan. Remove the duplicate private
copy from sba.rs and redirect player_has_cant_lose to the shared version.

Addresses gemini-code-assist review comment on sba.rs:267.
